evaluation/requirements.txt
pysqlite3==0.5.2
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
nltk==3.8.1
